requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0
python-telegram-bot==13.15
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.0
alembic==1.12.1 